Add - Modify
Modified the '/controllers/pagescontroller.php', and '/pages/welcome.blade.php' so the frontend for the blog, the index page posts list now pulls posts from the database.

Modified the '/posts/show.blade.php' with a button in the right sidebar to display '<< See All Posts' and directs to '/posts/index.blade.php'.

Modified the '/errors/404.blade.php' with a bootstrap jumbotran to look a little nicer than it was.

Added the '/errors/503.blade.php' file.